src/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.4.0
backports.asyncio.runner==1.2.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.13.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
gigachat==0.1.43
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
loguru==0.7.2
magic-filter==1.0.12
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.51.0
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tomli==2.3.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
win32_setctime==1.2.0
yarl==1.22.0
